RPF package modifications. Slight BasicMapPanel alteration.
* RPF modifications.  Changed RpfProductInfo to be an enum instead of a class, moved all defining information to it instead of being listed in the RpfConstants class.  Overhauled the RpfCoverage tool used by RpfLayer, simplified it and the RpfCoverageManager.	Updated RpfFrameProvider to use List instead of Vector.  Udpated classes to use doubles instead of floats for coordinates.* Modified BasicMapPanel startup order to change when startup projection settings are read and set on MapBean.

+++ src/corba/com/bbn/openmap/layer/rpf/corba/CRFPClient.java
@@ -231,7 +231,8 @@ public class CRFPClient
      * @param lrlon SE longitude
      * @param p a CADRG projection
      */
-    public Vector getCoverage(float ullat, float ullon, float lrlat, float lrlon, Projection p) {
+    public Vector getCoverage(double ullat, double ullon, double lrlat, double lrlon,
+                              Projection p) {
 
         CRFPCoverageBox[] boxes;
         Server serv = getServer();
@@ -270,7 +271,8 @@ public class CRFPClient
      * @param lrlon SE longitude
      * @param p a CADRG projection
      */
-    public Vector getCatalogCoverage(float ullat, float ullon, float lrlat, float lrlon, Projection p, String chartSeriesCode) {
+    public Vector getCatalogCoverage(double ullat, double ullon, double lrlat, double lrlon,
+                                     Projection p, String chartSeriesCode) {
         CRFPCoverageBox[] boxes;
         Server serv = getServer();
 
@@ -304,7 +306,8 @@ public class CRFPClient
      * @see #getCatalogCoverage(float ullat, float ullon, float lrlat, float
      *      lrlon, Projection p, String chartSeriesCode)
      */
-    public float getCalculatedCoverage(float ullat, float ullon, float lrlat, float lrlon, Projection p, String chartSeries) {
+    public float getCalculatedCoverage(double ullat, double ullon, double lrlat, double lrlon,
+                                       Projection p, String chartSeries) {
         if (chartSeries.equalsIgnoreCase(RpfViewAttributes.ANY)) {
             return 0f;
         }
